pyproject.toml
[tool.poetry]
name = "api-flask-dio-bank"
version = "0.1.0"
description = ""
authors = ["AlecLemos <aleclemos5@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "*"
flask-sqlalchemy = "*"
flask-migrate = "*"
flask-jwt-extended = "*"
pytest = "*"
pytest-mock = "*"

[tool.pytest.ini_options]
addopts = "-ra -vvv"
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
